+++ packages/gasket-plugin-logger/package.json
@@ -42,6 +42,7 @@
     "cross-env": "^7.0.3",
     "eslint": "^8.56.0",
     "eslint-config-godaddy": "^7.1.1",
+    "eslint-config-godaddy-typescript": "^4.0.3",
     "eslint-plugin-jest": "^28.6.0",
     "eslint-plugin-unicorn": "^55.0.0",
     "jest": "^29.7.0",
@@ -59,6 +60,16 @@
     ],
     "rules": {
       "unicorn/filename-case": "error"
-    }
+    },
+    "overrides": [
+      {
+        "files": [
+          "*.ts"
+        ],
+        "extends": [
+          "godaddy-typescript"
+        ]
+      }
+    ]
   }
 }
